move the changes I made from tl_detector.py to tl_classifier.py
